% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{ads.patch}
\alias{ads.patch}
\title{Updates an existing ad. This method supports patch semantics.}
\usage{
ads.patch(Ad, profileId, id)
}
\arguments{
\item{Ad}{The \link{Ad} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{id}{Ad ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}

Other Ad functions: \code{\link{Ad}},
  \code{\link{ads.insert}}, \code{\link{ads.update}}
}

