% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{advertiserGroups.insert}
\alias{advertiserGroups.insert}
\title{Inserts a new advertiser group.}
\usage{
advertiserGroups.insert(AdvertiserGroup, profileId)
}
\arguments{
\item{AdvertiserGroup}{The \link{AdvertiserGroup} object to pass to this method}

\item{profileId}{User profile ID associated with this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}

Other AdvertiserGroup functions: \code{\link{AdvertiserGroup}},
  \code{\link{advertiserGroups.patch}},
  \code{\link{advertiserGroups.update}}
}

