% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AdvertiserGroupsListResponse}
\alias{AdvertiserGroupsListResponse}
\title{AdvertiserGroupsListResponse Object}
\usage{
AdvertiserGroupsListResponse(advertiserGroups = NULL, nextPageToken = NULL)
}
\arguments{
\item{advertiserGroups}{Advertiser group collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
AdvertiserGroupsListResponse object
}
\description{
AdvertiserGroupsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Advertiser Group List Response
}

