% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ClickThroughUrlSuffixProperties}
\alias{ClickThroughUrlSuffixProperties}
\title{ClickThroughUrlSuffixProperties Object}
\usage{
ClickThroughUrlSuffixProperties(clickThroughUrlSuffix = NULL,
  overrideInheritedSuffix = NULL)
}
\arguments{
\item{clickThroughUrlSuffix}{Click-through URL suffix to apply to all ads in this entity's scope}

\item{overrideInheritedSuffix}{Whether this entity should override the inherited click-through URL suffix with its own defined value}
}
\value{
ClickThroughUrlSuffixProperties object
}
\description{
ClickThroughUrlSuffixProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Click Through URL Suffix settings.
}

