% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CompatibleFields}
\alias{CompatibleFields}
\title{CompatibleFields Object}
\usage{
CompatibleFields(crossDimensionReachReportCompatibleFields = NULL,
  floodlightReportCompatibleFields = NULL,
  pathToConversionReportCompatibleFields = NULL,
  reachReportCompatibleFields = NULL, reportCompatibleFields = NULL)
}
\arguments{
\item{crossDimensionReachReportCompatibleFields}{Contains items that are compatible to be selected for a report of type 'CROSS_DIMENSION_REACH'}

\item{floodlightReportCompatibleFields}{Contains items that are compatible to be selected for a report of type 'FLOODLIGHT'}

\item{pathToConversionReportCompatibleFields}{Contains items that are compatible to be selected for a report of type 'PATH_TO_CONVERSION'}

\item{reachReportCompatibleFields}{Contains items that are compatible to be selected for a report of type 'REACH'}

\item{reportCompatibleFields}{Contains items that are compatible to be selected for a report of type 'STANDARD'}
}
\value{
CompatibleFields object
}
\description{
CompatibleFields Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a response to the queryCompatibleFields method.
}

