% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ContentCategoriesListResponse}
\alias{ContentCategoriesListResponse}
\title{ContentCategoriesListResponse Object}
\usage{
ContentCategoriesListResponse(contentCategories = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{contentCategories}{Content category collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
ContentCategoriesListResponse object
}
\description{
ContentCategoriesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Content Category List Response
}

