% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{creativeAssets.insert}
\alias{creativeAssets.insert}
\title{Inserts a new creative asset.}
\usage{
creativeAssets.insert(CreativeAssetMetadata, profileId, advertiserId)
}
\arguments{
\item{CreativeAssetMetadata}{The \link{CreativeAssetMetadata} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{advertiserId}{Advertiser ID of this creative}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}

Other CreativeAssetMetadata functions: \code{\link{CreativeAssetMetadata}}
}

