% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CreativeGroup}
\alias{CreativeGroup}
\title{CreativeGroup Object}
\usage{
CreativeGroup(accountId = NULL, advertiserId = NULL,
  advertiserIdDimensionValue = NULL, groupNumber = NULL, id = NULL,
  name = NULL, subaccountId = NULL)
}
\arguments{
\item{accountId}{Account ID of this creative group}

\item{advertiserId}{Advertiser ID of this creative group}

\item{advertiserIdDimensionValue}{Dimension value for the ID of the advertiser}

\item{groupNumber}{Subgroup of the creative group}

\item{id}{ID of this creative group}

\item{name}{Name of this creative group}

\item{subaccountId}{Subaccount ID of this creative group}
}
\value{
CreativeGroup object
}
\description{
CreativeGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a creative group.
}
\seealso{
Other CreativeGroup functions: \code{\link{creativeGroups.insert}},
  \code{\link{creativeGroups.patch}},
  \code{\link{creativeGroups.update}}
}

