% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DfpSettings}
\alias{DfpSettings}
\title{DfpSettings Object}
\usage{
DfpSettings(dfp_network_code = NULL, dfp_network_name = NULL,
  programmaticPlacementAccepted = NULL, pubPaidPlacementAccepted = NULL,
  publisherPortalOnly = NULL)
}
\arguments{
\item{dfp_network_code}{DFP network code for this directory site}

\item{dfp_network_name}{DFP network name for this directory site}

\item{programmaticPlacementAccepted}{Whether this directory site accepts programmatic placements}

\item{pubPaidPlacementAccepted}{Whether this directory site accepts publisher-paid tags}

\item{publisherPortalOnly}{Whether this directory site is available only via DoubleClick Publisher Portal}
}
\value{
DfpSettings object
}
\description{
DfpSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
DFP Settings
}

