% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{File}
\alias{File}
\title{File Object}
\usage{
File(File.urls = NULL, dateRange = NULL, etag = NULL, fileName = NULL,
  format = NULL, id = NULL, lastModifiedTime = NULL, reportId = NULL,
  status = NULL, urls = NULL)
}
\arguments{
\item{File.urls}{The \link{File.urls} object or list of objects}

\item{dateRange}{The date range for which the file has report data}

\item{etag}{The eTag of this response for caching purposes}

\item{fileName}{The filename of the file}

\item{format}{The output format of the report}

\item{id}{The unique ID of this report file}

\item{lastModifiedTime}{The timestamp in milliseconds since epoch when this file was last modified}

\item{reportId}{The ID of the report this file was generated from}

\item{status}{The status of the report file}

\item{urls}{The URLs where the completed report file can be downloaded}
}
\value{
File object
}
\description{
File Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a File resource. A file contains the metadata for a report run. It shows the status of the run and holds the URLs to the generated report data if the run is finished and the status is 'REPORT_AVAILABLE'.
}
\seealso{
Other File functions: \code{\link{File.urls}}
}

