% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FloodlightConfigurationsListResponse}
\alias{FloodlightConfigurationsListResponse}
\title{FloodlightConfigurationsListResponse Object}
\usage{
FloodlightConfigurationsListResponse(floodlightConfigurations = NULL)
}
\arguments{
\item{floodlightConfigurations}{Floodlight configuration collection}
}
\value{
FloodlightConfigurationsListResponse object
}
\description{
FloodlightConfigurationsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Floodlight Configuration List Response
}

