% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{MetrosListResponse}
\alias{MetrosListResponse}
\title{MetrosListResponse Object}
\usage{
MetrosListResponse(metros = NULL)
}
\arguments{
\item{metros}{Metro collection}
}
\value{
MetrosListResponse object
}
\description{
MetrosListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metro List Response
}

