% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{OrderDocumentsListResponse}
\alias{OrderDocumentsListResponse}
\title{OrderDocumentsListResponse Object}
\usage{
OrderDocumentsListResponse(nextPageToken = NULL, orderDocuments = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{orderDocuments}{Order document collection}
}
\value{
OrderDocumentsListResponse object
}
\description{
OrderDocumentsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Order document List Response
}

