% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ProjectsListResponse}
\alias{ProjectsListResponse}
\title{ProjectsListResponse Object}
\usage{
ProjectsListResponse(nextPageToken = NULL, projects = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{projects}{Project collection}
}
\value{
ProjectsListResponse object
}
\description{
ProjectsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Project List Response
}

