% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Report.schedule}
\alias{Report.schedule}
\title{Report.schedule Object}
\usage{
Report.schedule(active = NULL, every = NULL, expirationDate = NULL,
  repeats = NULL, repeatsOnWeekDays = NULL, runsOnDayOfMonth = NULL,
  startDate = NULL)
}
\arguments{
\item{active}{Whether the schedule is active or not}

\item{every}{Defines every how many days, weeks or months the report should be run}

\item{expirationDate}{The expiration date when the scheduled report stops running}

\item{repeats}{The interval for which the report is repeated}

\item{repeatsOnWeekDays}{List of week days 'WEEKLY' on which scheduled reports should run}

\item{runsOnDayOfMonth}{Enum to define for 'MONTHLY' scheduled reports whether reports should be repeated on the same day of the month as 'startDate' or the same day of the week of the month}

\item{startDate}{Start date of date range for which scheduled reports should be run}
}
\value{
Report.schedule object
}
\description{
Report.schedule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The report's schedule. Can only be set if the report's 'dateRange' is a relative date range and the relative date range is not 'TODAY'.
}
\seealso{
Other Report functions: \code{\link{Report.criteria}},
  \code{\link{Report.crossDimensionReachCriteria}},
  \code{\link{Report.delivery}},
  \code{\link{Report.floodlightCriteria.reportProperties}},
  \code{\link{Report.floodlightCriteria}},
  \code{\link{Report.pathToConversionCriteria.reportProperties}},
  \code{\link{Report.pathToConversionCriteria}},
  \code{\link{Report.reachCriteria}}, \code{\link{Report}},
  \code{\link{reports.insert}},
  \code{\link{reports.patch}}, \code{\link{reports.update}}
}

