% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{SiteSettings}
\alias{SiteSettings}
\title{SiteSettings Object}
\usage{
SiteSettings(activeViewOptOut = NULL, creativeSettings = NULL,
  disableBrandSafeAds = NULL, disableNewCookie = NULL,
  lookbackConfiguration = NULL, tagSetting = NULL,
  videoActiveViewOptOut = NULL)
}
\arguments{
\item{activeViewOptOut}{Whether active view creatives are disabled for this site}

\item{creativeSettings}{Site-wide creative settings}

\item{disableBrandSafeAds}{Whether brand safe ads are disabled for this site}

\item{disableNewCookie}{Whether new cookies are disabled for this site}

\item{lookbackConfiguration}{Lookback window settings for this site}

\item{tagSetting}{Configuration settings for dynamic and image floodlight tags}

\item{videoActiveViewOptOut}{Whether Verification and ActiveView are disabled for in-stream video creatives on this site}
}
\value{
SiteSettings object
}
\description{
SiteSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Site Settings
}

