% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{SizesListResponse}
\alias{SizesListResponse}
\title{SizesListResponse Object}
\usage{
SizesListResponse(sizes = NULL)
}
\arguments{
\item{sizes}{Size collection}
}
\value{
SizesListResponse object
}
\description{
SizesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Size List Response
}

