% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Subaccount}
\alias{Subaccount}
\title{Subaccount Object}
\usage{
Subaccount(accountId = NULL, availablePermissionIds = NULL, id = NULL,
  name = NULL)
}
\arguments{
\item{accountId}{ID of the account that contains this subaccount}

\item{availablePermissionIds}{IDs of the available user role permissions for this subaccount}

\item{id}{ID of this subaccount}

\item{name}{Name of this subaccount}
}
\value{
Subaccount object
}
\description{
Subaccount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a DCM subaccount.
}
\seealso{
Other Subaccount functions: \code{\link{subaccounts.insert}},
  \code{\link{subaccounts.patch}},
  \code{\link{subaccounts.update}}
}

